Use shared hook with CI guard
