+++ src/applications/mhv/landing-page/containers/App.jsx
@@ -6,7 +6,7 @@ import { RequiredLoginView } from '@department-of-veterans-affairs/platform-user
 
 import LandingPage from '../components/LandingPage';
 import { resolveLandingPageLinks } from '../utilities/data';
-import { useDatadogRum } from '../hooks/useDatadogRum';
+import { useDatadogRum } from '../../shared/hooks/useDatadogRum';
 import {
   isAuthenticatedWithSSOe,
   isLandingPageEnabledForUser,
@@ -32,7 +32,20 @@ const App = () => {
     [featureToggles, ssoe],
   );
 
-  useDatadogRum();
+  const datadogRumConfig = {
+    applicationId: '1f81f762-c3fc-48c1-89d5-09d9236e340d',
+    clientToken: 'pub3e48a5b97661792510e69581b3b272d1',
+    site: 'ddog-gov.com',
+    service: 'mhv-on-va.gov',
+    sessionSampleRate: 100,
+    sessionReplaySampleRate: 10,
+    trackInteractions: true,
+    trackUserInteractions: true,
+    trackResources: true,
+    trackLongTasks: true,
+    defaultPrivacyLevel: 'mask-user-input',
+  };
+  useDatadogRum(datadogRumConfig);
 
   const loading =
     vamcEhrData.loading || featureToggles.loading || profile.loading;
