--- src/applications/mhv/secure-messaging/hooks/useDatadogRum.jsx
@@ -1,38 +1,0 @@
-import { useEffect } from 'react';
-import { datadogRum } from '@datadog/browser-rum';
-
-import environment from '@department-of-veterans-affairs/platform-utilities/environment';
-
-const initializeDatadogRum = () => {
-  if (
-    // Prevent RUM from running on local/CI environments.
-    !environment.isLocalhost() &&
-    // Prevent re-initializing the SDK.
-    !window.DD_RUM?.getInitConfiguration() &&
-    !window.Mocha
-  ) {
-    datadogRum.init({
-      applicationId: '02c72297-5059-4ed8-8472-874276f4a9b2',
-      clientToken: 'pub1325dfe255119729611410e2f47f4f99',
-      site: 'ddog-gov.com',
-      service: 'va.gov-mhv-secure-messaging',
-      env: environment.vspEnvironment(),
-      sessionSampleRate: 100, // controls the percentage of overall sessions being tracked
-      sessionReplaySampleRate: 50, // is applied after the overall sample rate, and controls the percentage of sessions tracked as Browser RUM & Session Replay
-      trackInteractions: true,
-      trackUserInteractions: true,
-      trackResources: true,
-      trackLongTasks: true,
-      defaultPrivacyLevel: 'mask-user-input',
-    });
-    datadogRum.startSessionReplayRecording();
-  }
-};
-
-const useDatadogRum = () => {
-  useEffect(() => {
-    initializeDatadogRum();
-  }, []);
-};
-
-export { useDatadogRum };
